requirements.txt
python-telegram-bot
requests
amazon-paapi